refactor: represent all money values as integer cents
Cents are the universal unit throughout the application. Two boundary
conversions exist in src/money.py:

  dollars_to_cents — called in pipeline handlers when persisting
    LLM-extracted dollar floats from ExtractionResult to the DB
  cents_to_dollars — called in the SMS formatter when building
    user-facing reply text from stored cent values

All internal types updated: Job.pay_rate, WorkGoal.target_earnings,
JobCandidate.earnings, and MatchResult.total_earnings are now int.
_dp_select works natively in integer cents — SCALE=100 removed.
Migration 006 converts existing float columns with ROUND(col*100).
Test fixtures accept dollar amounts and convert to cents internally.

+++ src/jobs/schemas.py
@@ -8,7 +8,7 @@ class JobCreate(BaseModel):
     message_id: int = Field(gt=0)
     description: str = Field(min_length=1, max_length=1000)
     location: str = Field(min_length=1, max_length=255)
-    pay_rate: Optional[float] = Field(default=None, gt=0)
+    pay_rate: Optional[int] = Field(default=None, gt=0)
     pay_type: str = Field(default="unknown", min_length=1, max_length=50)
     estimated_duration_hours: Optional[float] = Field(default=None, gt=0)
     raw_duration_text: Optional[str] = Field(default=None, max_length=255)
